+++ README.md
@@ -3,12 +3,14 @@
 This is a framework for building, training and analyzing neural networks created using NumPy only. Similar to PyTorch I introduced a `Tensor`-object as the central building block that keeps track of its data, gradients and more. However, this framework does not support autograd. Gradients are computed within a network's layer.
 
 ```Python
-a = Tensor([1, 2, 3])
-b = nn.randn((3,))
+import walnut
+
+a = walnut.Tensor([1, 2, 3])
+b = walnut.randn((3,))
 
 c = a + b
 
-d = nn.tensor.zeros((5, 5))
+d = walnut.zeros((5, 5))
 ```
 
 ### Data preprocessing
@@ -20,23 +22,25 @@ The framework offers some utility functions, such as `split_train_val_test()` to
 Models can be built using a variety of layers, including trainable layers such as linear or convolutional layers. Most of the common activation functions as well as normalization can be applied.
 
 ```python
-model = networks.Sequential(input_shape=(4,), mdl_layers=[
-    layers.Linear(out_channels=16, act_fn=activations.Tanh(), init_fn=inits.Kaiming),
-    layers.Linear(out_channels=16, act_fn=activations.Tanh(), init_fn=inits.Kaiming),
-    layers.Linear(out_channels=16, act_fn=activations.Tanh(), init_fn=inits.Kaiming),
-    layers.Linear(out_channels=3, act_fn=activations.Softmax(), init_fn=inits.Kaiming)
+import walnut.nn as nn
+
+model = nn.Sequential(input_shape=(4,), mdl_layers=[
+    nn.layers.Linear(out_channels=16, act_fn=nn.activations.Tanh(), init_fn=nn.inits.Kaiming),
+    nn.layers.Linear(out_channels=16, act_fn=nn.activations.Tanh(), init_fn=nn.inits.Kaiming),
+    nn.layers.Linear(out_channels=16, act_fn=nn.activations.Tanh(), init_fn=nn.inits.Kaiming),
+    nn.layers.Linear(out_channels=3, act_fn=nn.activations.Softmax(), init_fn=nn.inits.Kaiming)
 ])
 ```
 
 ```python
-model = networks.Sequential(input_shape=(1, 28, 28), mdl_layers=[
-    layers.Convolution(out_channels=8, kernel_shape=(3, 3), act_fn=activations.Relu(), norm_fn=norms.Layernorm()),
-    layers.MaxPooling(p_window=(2, 2)),
-    layers.Convolution(out_channels=16, kernel_shape=(3, 3), act_fn=activations.Relu(), norm_fn=norms.Layernorm()),
-    layers.MaxPooling(p_window=(2, 2)),
-    layers.Flatten(),
-    layers.Linear(out_channels=64, act_fn=activations.Relu(), norm_fn=norms.Layernorm()),
-    layers.Linear(out_channels=10, act_fn=activations.Softmax())
+model = nn.Sequential(input_shape=(1, 28, 28), mdl_layers=[
+    nn.layers.Convolution(out_channels=8, kernel_shape=(3, 3), act_fn=nn.activations.Relu(), norm_fn=nn.norms.Layernorm()),
+    nn.layers.MaxPooling(p_window=(2, 2)),
+    nn.layers.Convolution(out_channels=16, kernel_shape=(3, 3), act_fn=nn.activations.Relu(), norm_fn=nn.norms.Layernorm()),
+    nn.layers.MaxPooling(p_window=(2, 2)),
+    nn.layers.Flatten(),
+    nn.layers.Linear(out_channels=64, act_fn=nn.activations.Relu(), norm_fn=nn.norms.Layernorm()),
+    nn.layers.Linear(out_channels=10, act_fn=nn.activations.Softmax())
 ])
 ```
 
@@ -46,17 +50,17 @@ The model can be trained using common algorithms, such as SGD or Adam.
 
 ```python
 model.compile(
-    optimizer=optimizers.SGD(l_r=1e-2, momentum=0.9, nesterov=True),
-    loss_fn=losses.Crossentropy(),
-    metric=metrics.accuracy
+    optimizer=nn.optimizers.SGD(l_r=1e-2, momentum=0.9, nesterov=True),
+    loss_fn=nn.losses.Crossentropy(),
+    metric=nn.metrics.accuracy
 )
 ```
 
 ```python
 model.compile(
-    optimizer=optimizers.Adam(l_r=1e-3),
-    loss_fn=losses.Crossentropy(),
-    metric=metrics.accuracy
+    optimizer=nn.optimizers.Adam(l_r=1e-3),
+    loss_fn=nn.losses.Crossentropy(),
+    metric=nn.metrics.accuracy
 )
 ```
 
@@ -78,7 +82,7 @@ The framework also provides some functions for analyzing a models' parameters an
 
 All layers can also be used individually without a model and their parameters can be inspected.
 ```python
-conv = layers.Convolution(out_channels=2)
+conv = nn.layers.Convolution(out_channels=2)
 
 conv.x = X
 conv.w = W
