--- numpynn/__init__.py
@@ -1,6 +1,0 @@
-"""numpynn module"""
-
-from numpynn import activations, inits, layers, losses, metrics, networks, norms, optimizers, paddings, preprocessing, tensor, utils
-from numpynn.tensor import *
-
-utils.set_numpy_format()
