+++ walnut/__init__.py
@@ -1,0 +1,7 @@
+"""numpynn module"""
+
+import walnut.nn as nn
+from walnut import preprocessing, tensor, utils
+from walnut.tensor import *
+
+utils.set_numpy_format()
