+++ walnut/nn/__init__.py
@@ -1,0 +1,4 @@
+"""numpynn neural network module"""
+
+from walnut.nn import activations, inits, layers, losses, metrics, networks, norms, optimizers, paddings
+from walnut.nn.networks import Sequential
